Move "Create new collection" out of scroll content
